Learn Driver Android app

+++ app/src/main/java/com/example/sublearn/nearMeList.kt
@@ -1,6 +1,6 @@
 package com.example.sublearn
 
 //Create data class to parse .csv data of suburb name, suburb distance, and suburb crash statistic
-data class nearMeList(val text1: String, val text2: Double, val text3: Double) {
+data class nearMeList(val text1: String, val text2: Double, val text3: Int) {
 
 }
